Inline lightbox helpers
